+++ R/checkNames.R
@@ -1,7 +1,7 @@
 #' Check names to comply to specific rules
 #'
 #' @description
-#' Similar to \code{\link{checkNamed}} but you can pass the names directly.
+#' Performs various checks on character vectors, usually names.
 #'
 #' @templateVar fn Named
 #' @param x [\code{character} || \code{NULL}]\cr
@@ -13,6 +13,7 @@
 #'  \item{named:}{Checks \code{x} for regular names which excludes names to be \code{NA} or empty (\code{""}).}
 #'  \item{unique:}{Performs checks like with \dQuote{named} and additionally tests for non-duplicated names.}
 #'  \item{strict:}{Performs checks like with \dQuote{unique} and additionally fails for names with UTF-8 characters and names which do not comply to R's variable name restrictions. As regular expression, this is \dQuote{^[.]*[a-zA-Z]+[a-zA-Z0-9._]*$}.}
+#'  \item{ids:}{Same as \dQuote{strict}, but does not enforce uniqueness.}
 #'  }
 #'  Note that for zero-length \code{x}, all these name checks evaluate to \code{TRUE}.
 #' @param subset.of [\code{character}]\cr